Add supportable module & implementation
